Move cnl::shift into a separate header file
