+++ include/cnl/num_traits.h
@@ -16,6 +16,7 @@
 #include "bits/num_traits/digits.h"
 #include "bits/num_traits/from_value.h"
 #include "bits/num_traits/set_digits.h"
+#include "bits/num_traits/shift.h"
 #include "bits/power.h"
 #include "bits/type_traits.h"
 #include "bits/type_traits/is_integral.h"
@@ -133,53 +134,6 @@ namespace cnl {
                 _impl::enable_if_t<_impl::are_composite<Args ...>::value, int> dummy = 0>
         constexpr Result for_rep(F f, Args &&...args) {
             return for_rep<Result>(f, to_rep(std::forward<Args>(args))...);
-        }
-    }
-
-    ////////////////////////////////////////////////////////////////////////////////
-    // cnl::shift
-
-    template<int Digits, int Radix, class S, class Enable = void>
-    struct shift;
-
-    namespace _impl {
-        // fundamental integer-friendly cnl::shift algorithm
-        template<int Digits, int Radix, typename S, class Enable = void>
-        struct default_shift;
-
-        template<int Digits, int Radix, typename S>
-        struct default_shift<Digits, Radix, S, _impl::enable_if_t<0<=Digits>> {
-            constexpr auto operator()(S const& s) const
-            -> decltype(s*power<S, Digits, Radix>())
-            {
-                return s*power<S, Digits, Radix>();
-            }
-        };
-
-        // cnl::default_shift<-ve, cnl::constant<>>
-        template<int Digits, int Radix, typename S>
-        struct default_shift<Digits, Radix, S, _impl::enable_if_t<Digits<0>> {
-            constexpr auto operator()(S const& s) const
-            -> decltype(s/power<S, -Digits, Radix>())
-            {
-                return s/power<S, -Digits, Radix>();
-            }
-        };
-    }
-
-    // cnl::shift<..., fundamental-integer>
-    template<int Digits, int Radix, class S>
-    struct shift<Digits, Radix, S, _impl::enable_if_t<cnl::is_integral<S>::value>>
-            : _impl::default_shift<Digits, Radix, S> {
-    };
-
-    namespace _impl {
-        // cnl::_impl::shift - convenience wrapper for cnl::shift
-        template<int Digits, int Radix=2, class S>
-        constexpr auto shift(S const& s)
-        -> decltype(cnl::shift<Digits, Radix, S>{}(s))
-        {
-            return cnl::shift<Digits, Radix, S>{}(s);
         }
     }
